Clean up DAP 2 constraint parser package
