+++ dap2-service/src/main/scala/latis/util/dap2/parser/ConstraintParser.scala
@@ -1,7 +1,9 @@
-package latis.util.dap2
+package latis.util.dap2.parser
 
 import atto._, Atto._
 import cats.implicits._
+
+import ast._
 
 /**
  * Module for parsing DAP 2 constraint expressions.
@@ -10,7 +12,6 @@ import cats.implicits._
  * follow the DAP 2 specification.
  */
 object ConstraintParser {
-  import internal._
 
   /**
    * Parse a DAP 2 constraint expression into a sequence of LaTiS
@@ -30,76 +31,6 @@ object ConstraintParser {
       Right(ast.ConstraintExpression(List.empty))
     }
   }
-}
-
-/**
- * Module for the DAP 2 constraint expression AST.
- */
-object ast {
-
-  sealed abstract trait CExpr
-
-  final case class Projection(
-    names: List[String]
-  ) extends CExpr
-
-  final case class Selection(
-    name: String, op: SelectionOp, value: String
-  ) extends CExpr
-
-  final case class Operation(
-    name: String, args: List[String]
-  ) extends CExpr
-
-  sealed abstract trait SelectionOp
-  final case object Gt extends SelectionOp
-  final case object Lt extends SelectionOp
-  final case object Eq extends SelectionOp
-  final case object GtEq extends SelectionOp
-  final case object LtEq extends SelectionOp
-  final case object EqEq extends SelectionOp
-  final case object NeEq extends SelectionOp
-  final case object Tilde extends SelectionOp
-  final case object EqTilde extends SelectionOp
-  final case object NeEqTilde extends SelectionOp
-
-  /**
-   * Wrapper for DAP 2 constraint expressions.
-   */
-  final case class ConstraintExpression(exprs: List[CExpr]) extends AnyVal
-
-  /**
-   * Pretty print a constraint expression.
-   *
-   * @param expr constraint expression to pretty print
-   */
-  def pretty(expr: ConstraintExpression): String = {
-    def prettyOp(op: SelectionOp): String = op match {
-      case Gt        => ">"
-      case Lt        => "<"
-      case Eq        => "="
-      case GtEq      => ">="
-      case LtEq      => "<="
-      case EqEq      => "=="
-      case NeEq      => "!="
-      case Tilde     => "~"
-      case EqTilde   => "=~"
-      case NeEqTilde => "!=~"
-    }
-
-    expr.exprs.map {
-      case Projection(n)       => n.mkString(",")
-      case Selection(n, op, v) => s"$n${prettyOp(op)}$v"
-      case Operation(n, args)  => s"$n(${args.mkString(",")})"
-    }.mkString("&")
-  }
-}
-
-/**
- * Parsers used within [[ConstraintParser]].
- */
-object internal {
-  import ast._
 
   /*
    * A note about "|" and "choice":
@@ -112,13 +43,13 @@ object internal {
    * succeed.
    */
 
-  def subexpression: Parser[CExpr] =
+  private def subexpression: Parser[CExpr] =
     selection | operation | projection
 
-  def projection: Parser[CExpr] =
+  private def projection: Parser[CExpr] =
     sepBy1(variable.token, char(',').token).map(xs => Projection(xs.toList))
 
-  def selectionOp: Parser[SelectionOp] = choice(
+  private def selectionOp: Parser[SelectionOp] = choice(
     string("~")   ~> ok(Tilde),
     string(">=")  ~> ok(GtEq),
     string("<=")  ~> ok(LtEq),
@@ -131,38 +62,38 @@ object internal {
     string("=")   ~> ok(Eq)
   )
 
-  def selection: Parser[CExpr] = for {
+  private def selection: Parser[CExpr] = for {
     name  <- variable.token
     op    <- selectionOp.token
     value <- time | number | stringLit
   } yield Selection(name, op, value)
 
-  def operation: Parser[CExpr] = for {
+  private def operation: Parser[CExpr] = for {
     name <- identifier
     argP  = time | number | stringLit | variable
     args <- parens(sepBy(argP.token, char(',').token))
   } yield Operation(name, args)
 
-  def variable: Parser[String] =
+  private def variable: Parser[String] =
     sepBy1(identifier, char('.')).map(_.toList.mkString("."))
 
-  def identifier: Parser[String] = for {
+  private def identifier: Parser[String] = for {
     init <- letter | char('_')
     rest <- many(letterOrDigit | char('_'))
   } yield (init :: rest).mkString
 
-  def stringLit: Parser[String] = for {
+  private def stringLit: Parser[String] = for {
     lit <- stringLiteral
   } yield "\"" + lit + "\""
 
-  def sign: Parser[String] = string("+") | string("-")
+  private def sign: Parser[String] = string("+") | string("-")
 
-  def integer: Parser[String] = for {
+  private def integer: Parser[String] = for {
     s <- sign | ok("")
     n <- stringOf1(digit)
   } yield s + n
 
-  def decimal: Parser[String] = {
+  private def decimal: Parser[String] = {
     // A decimal variant for which the fractional part is optional.
     val n1: Parser[String] = for {
       int  <- stringOf1(digit)
@@ -183,15 +114,15 @@ object internal {
     } yield s + n
   }
 
-  def scientific: Parser[String] = for {
+  private def scientific: Parser[String] = for {
     significand <- decimal | integer
     e           <- string("e") | string("E")
     exponent    <- integer
   } yield significand + e + exponent
 
-  def number: Parser[String] = scientific | decimal | integer
+  private def number: Parser[String] = scientific | decimal | integer
 
-  def time: Parser[String] = {
+  private def time: Parser[String] = {
     def stringN(n: Int, p: Parser[Char]): Parser[String] =
       count(n, p).map(_.mkString)
