+++ dap2-service/src/test/scala/latis/util/dap2/parser/ConstraintParserProps.scala
@@ -1,4 +1,4 @@
-package latis.util.dap2
+package latis.util.dap2.parser
 
 import org.scalacheck._
 import org.scalacheck.ScalacheckShapeless._
